% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-buildClass.R
\docType{methods}
\name{buildClass}
\alias{buildClass}
\alias{buildClass,RS,character-method}
\title{Build Extended S4 Class}
\usage{
buildClass(object, method, freq, ind, lagMax, verify, parallel)

\S4method{buildClass}{RS,character}(object, method, freq, ind, lagMax,
  verify, parallel)
}
\arguments{
\item{object}{An object of \code{\link{RS}} class.}

\item{method}{An character value indicating the method to apply. Available method in this package are:
\itemize{
 \item{dlm: }{build a \code{\link{SM.dlm}} class;}
 \item{hl: }{build a \code{\link{SM.HL}} class.}
}}

\item{freq}{Object with type ANY. \itemize{
 \item{dlm method: }{a positive intervalue indicating the seasonality of the time series' structure;}
 \item{hl method: }{a NULL value, it is not necessary for \code{\link{SM.HL}} class.}
}}

\item{ind}{R object with type ANY. \itemize{
 \item{dlm method: }{a list containing outliers or human intervention indexes;}
 \item{hl method: }{a numeric value indicating the offset value.}
}}

\item{lagMax}{R object with type ANY. For dlm method, \code{lagMax} is an integer value indicating
the maximum effect lag for each outlier. Check section Details in \code{\link{SM.dlm}}
for more information.}

\item{verify}{R object with type ANY. For dlm method, \code{verify} is a logical value indicating
whether the validation process should be applied. Check section Details in \code{\link{SM.dlm}}
for more information.}

\item{parallel}{R object with type ANY. For dlm method, \code{verify} is a logical value indicating
whether the parallel process should be applied.}
}
\description{
\code{buildClass} extends the class \code{\link{RS}} to \code{\link{SM.dlm}} and \code{\link{SM.HL}}
 by adding the additional informations.
}
\examples{
n <- 50
# An example to create the RS class.
my_first_class <- RS(sigb = rnorm(n, mean = 0, sd = 1),
                     epsb = rnorm(n, mean = 0, sd = 2),
                     epsp = rnorm(n, mean = 1, sd = 1))

# build class
## build dlm model
dlm_class <- buildClass(object = my_first_class,
                        method = "dlm", freq = 1,
                        ind = list(), lagMax = 3,
                        verify = TRUE, parallel = TRUE)
## build HL model
hl_class <- buildClass(object = my_first_class,
                       method = "HL", ind = 9.5)
### ind in this case is offset
}
