% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-[.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,RS,character,missing,missing-method}
\alias{[,SM.dlm,character,missing,missing-method}
\alias{[,SM.dlm.fitted,character,missing,missing-method}
\alias{[,SM.HL,character,missing,missing-method}
\alias{[,SM.rnn,character,missing,missing-method}
\title{Extract Slots of an S4 Class}
\usage{
\S4method{[}{RS,character,missing,missing}(x, i, j, drop)

\S4method{[}{SM.dlm,character,missing,missing}(x, i, j, drop)

\S4method{[}{SM.dlm.fitted,character,missing,missing}(x, i, j, drop)

\S4method{[}{SM.HL,character,missing,missing}(x, i, j, drop)

\S4method{[}{SM.rnn,character,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{An S4 object from which to extact slot's values.}

\item{i}{A character value indicating the slot's name.}

\item{j}{Missing value which is not used for objects from this package.}

\item{drop}{Same as \code{j}.}
}
\description{
Operators acting on existing classes to extract slot's values.
}
\examples{
n <- 50
# An example to create the RS class.
my_first_class <- RS(sigb = rnorm(n, mean = 0, sd = 1),
                     epsb = rnorm(n, mean = 0, sd = 2),
                     epsp = rnorm(n, mean = 1, sd = 1))
# Extract the slot "sigb"
sigb <- my_first_class["sigb"]

}
