% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-extractMeasures.R
\docType{methods}
\name{extractMeasures}
\alias{extractMeasures}
\alias{extractMeasures,SM.dlm.fitted,logical-method}
\alias{extractMeasures,SM.HL,logical-method}
\title{Extract and Plot Measurement}
\usage{
extractMeasures(object, plot)

\S4method{extractMeasures}{SM.dlm.fitted,logical}(object, plot)

\S4method{extractMeasures}{SM.HL,logical}(object, plot)
}
\arguments{
\item{object}{Object to be extracted the stored measurement.}

\item{plot}{A logical value indicating whether the graphic should be plotted.}
}
\description{
Extract and plot the electrical conductivity of pore water.
}
\note{
Check the section \code{Examples} in \code{\link{SM.dlm.fitted}} and \code{\link{SM.HL}}
for practical use in each case.
}
