% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,SM.dlm-method}
\title{Build and train the SM.dlm Class in Package sm4sd}
\usage{
fit(object)

\S4method{fit}{SM.dlm}(object)
}
\arguments{
\item{object}{An object of \code{\link{SM.dlm}}.}
}
\description{
Creating an object of \code{SM.dlm.fitted} class by build and train the dlm model based on the information
 of input.
}
\note{
Check the section \code{Examples} in \code{\link{SM.dlm.fitted}} for practical use.
}
