% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getMod.R
\docType{methods}
\name{getMod}
\alias{getMod}
\alias{getMod,SM.dlm.fitted,missing-method}
\title{Extract and/or Build the Corresponding Model}
\usage{
getMod(object, name)

\S4method{getMod}{SM.dlm.fitted,missing}(object)
}
\arguments{
\item{object}{An existing class in this package.}

\item{name}{Missing value, it's not defined for current version.}
}
\description{
Rebuilding the model based on the informations stored in the object.
}
\note{
Check the section \code{Examples} in \code{\link{SM.dlm.fitted}} for practical use.
}
