% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,RS,missing-method}
\title{Methods for Function Plot in Package sm4sd}
\usage{
\S4method{plot}{RS,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of \code{\link{RS}} class.}

\item{y}{Missing object.}

\item{...}{Additional graphical arguments, see plot, plot.default and par.}
}
\description{
Plot slots of \code{\link{RS}} class using "timeSeries" plot.
}
\examples{
n <- 50
# An example to create the RS class.
my_first_class <- RS(sigb = rnorm(n, mean = 0, sd = 1),
                     epsb = rnorm(n, mean = 0, sd = 2),
                     epsp = rnorm(n, mean = 1, sd = 1))

# plot the class as timeseries object
plot(my_first_class)

}
