% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-residualDiag.R
\docType{methods}
\name{residualDiag}
\alias{residualDiag}
\alias{residualDiag,SM.dlm.fitted,numeric-method}
\title{Residual Diagnostic for DLM Model}
\usage{
residualDiag(object, gamma)

\S4method{residualDiag}{SM.dlm.fitted,numeric}(object, gamma)
}
\arguments{
\item{object}{An object to be analyzed.}

\item{gamma}{Numeric value between 0 and 1 indicating the confidence level.}
}
\description{
Residual Diagnostic including several normal test and residual plots.
}
\section{Details}{

 The residual diagnostic include the confidence band plot, chisq test for normalized error plot,
 \code{\link[stats:qqplot]{Quantile-Quantile Plots}} and
 \code{\link[lmtest:dwtest]{Durbin-Waston Test for autocorrelation}}.
}

