% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-RNN_predict.R
\name{RNN_predict}
\alias{RNN_predict}
\title{Predict Function for SM.rnn Class}
\usage{
RNN_predict(object, sigb, epsp)
}
\arguments{
\item{object}{An SM.rnn class.}

\item{sigb, epsp}{A n-dimensional numeric vector of electrical conductivity of bulk soil measures and
electrical permitivity of pore water measures.}
}
\description{
Simulation of \ifelse{html}{\out{&#949<sub>b</sub>}}{\eqn{\epsilon_b}{ASCII}} based on the new
 \ifelse{html}{\out{&#949<sub>p</sub>}}{\eqn{\epsilon_p}{ASCII}} and
 \ifelse{html}{\out{&#963<sub>b</sub>}}{\eqn{\sigma_b}{ASCII}} according to the pattern captured by
 SM.rnn model.
}
\section{Note}{

Before the prediction is applied, the data will be preprocessed according to the preprocessed method
  defined in the SM.rnn class.
}

