% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,RS-method}
\alias{show,SM.dlm-method}
\alias{show,SM.dlm.fitted-method}
\alias{show,SM.HL-method}
\alias{show,SM.rnn-method}
\title{Show an Object}
\usage{
\S4method{show}{RS}(object)

\S4method{show}{SM.dlm}(object)

\S4method{show}{SM.dlm.fitted}(object)

\S4method{show}{SM.HL}(object)

\S4method{show}{SM.rnn}(object)
}
\arguments{
\item{object}{An s4 class.}
}
\description{
Display the object by printing according to the class object. Check "Details" section for more info
 in each case.
}
\section{Details}{

Simplying the print for each case:\itemize{
 \item{\code{\link{RS}}: }{first 25 elements for each slot;}
 \item{\code{\link{SM.dlm}}: }{call, seasonality of time series, index of event indicators, extra parameter as
  lagMax, verify and parallel, and inherited part of \code{\link{RS}} class.}
 \item{\code{\link{SM.dlm.fitted}}: } {call, best model parameters including dlm parameter (V, W, outlier
  ajustment), and associated lag, seasonality, index of event indicators and first 5 rows of tracking
  ordered decreasely by performance.}
 \item{\code{\link{SM.HL}}: } {call, corresponding offset and inherited part of \code{\link{RS}} class.}
}
}

