% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_classes.R
\docType{class}
\name{SM.dlm.fitted-class}
\alias{SM.dlm.fitted-class}
\alias{SM.dlm.fitted}
\title{Class SM.dlm.fitted --- An S4 class to represent the fitted dynamic linear model}
\description{
The SM.dlm.fitted class store information about fitted dynamic linear model and its pre-definition.
}
\section{Objects from the Class}{

SM.dlm.fitted is created by applying method \code{\link{fit}} to the \code{\link{SM.dlm}} class.
}

\section{Slots}{

  \describe{
    \item{\code{parameters}}{A list containing the best best parameters which are used to build the corresponded
      model by \code{\link{getMod}}.}
    \item{\code{filtered}}{A list contaning the filtered value. For more information, check the Value section of
      \code{\link[dlm]{dlmFilter}}.}
    \item{\code{smoothed}}{A list contaning the smoothed value. See also \code{\link[dlm]{dlmSmooth}}.}
    \item{\code{lags}}{A list contaning the best lag for each outlier indicator. For more detail, see the section
       Details of \code{\link{SM.dlm}}.}
    \item{\code{seasonalSign}}{A character indicating the significance of sesonal part in the model:\itemize{
        \item null: if the data doesn't have sesonal part (freq == 1)
        \item TRUE: the seasonality is significant in the model
        \item FALSE: the seasonality is not significant
        }. Check details for class building.}
    \item{\code{tracking}}{A data frame contaning the tracking of the gridsearch history ordered by the negative
      log likelihood.}
   }
}

\section{Methods}{

Available methods for this class are: \itemize{
 \item \code{\link{show}}
 \item \code{\link{print}}
 \item \code{\link{get}} (`[`): giving the value of a slot
 \item \code{\link{getMod}}: returning the dlm model
 \item \code{\link{residualDiag}}: residual diagnostic for model validation
 \item \code{\link{extractMeasures}}: measurement extraction and visualization.
}
}

\section{Extends}{

 From \code{\link{SM.dlm}}, directly.
}

\examples{
n <- 50
rs.class <- RS(sigb = rnorm(n, mean = 0, sd = 1),
               epsb = rnorm(n, mean = 0, sd = 2),
               epsp = rnorm(n, mean = 1, sd = 1))
# The inputs we created are random variable,
# then they have non-stationality structure.
freq <- 1
# Suppose there is a rain at index 1 and 20, and irrigation at 5.
index <- list(rain = c(1, 20), watering = c(5))
lagMax <- 10
verify <- parallel <- TRUE
lagMax <- 1

 dlm.class <- buildClass(object = rs.class, method = "dlm",
                        freq = freq, ind = index, lagMax = lagMax,
                        verify = verify, parallel = parallel)
 dlm.fitted.class <- fit(dlm.class)
 dlm_model <- getMod(dlm.fitted.class)

}
