% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_classes.R
\docType{class}
\name{SM.HL-class}
\alias{SM.HL-class}
\alias{SM.HL}
\title{Class SM.HL --- An S4 class to represent the Hilhorst model.}
\description{
Class SM.HL --- An S4 class to represent the Hilhorst model.
}
\section{Slots}{

\describe{
\item{\code{call}}{An object of class \code{\link[base]{call}} returning an unevaluated function call.}

\item{\code{offset}}{A no negative numeric value indicating
\ifelse{html}{\out{&#949<sub>&#963<sub>b</sub>=0</sub>}}{\eqn{\epsilon_{\sigma_b}=0}{ASCII}}.}
}}

\section{Note}{

 A hilhorst model is a deterministic model for this problem. For more detail, please check
  \code{\link{sm4sd Package}} and see also the paper of M.A.Hilhorst:
  \href{https://pdfs.semanticscholar.org/2ff3/70a503943086f073ee52f16b33c1b1ea5fc5.pdf}{A Pore Water Conductivity Sensor}.
}

