% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_classes.R
\docType{class}
\name{SM.rnn-class}
\alias{SM.rnn-class}
\alias{SM.rnn}
\title{Class SM.rnn --- An S4 class contains a recurrent neural network model for
 \ifelse{html}{\out{&#963<sub>b</sub>}}{\eqn{\sigma_b}{ASCII}} simulation}
\description{
SM.rnn class contains the trained rnn model and its parameters.
}
\section{objects from the Class}{

 Objects are created by the method \code{\link{buildRNN}}.
}

\section{Slots}{

 \describe{
   \item{\code{call}}{An object of class \code{\link[base]{call}} returning an unevaluated function call.}
   \item{\code{model_path}}{The directory where the trained rnn model should be stored.}
   \item{\code{compile_options}}{A list contanning compile options of keras model, check
    \code{\link{buildRNN}} for more detail.}
   \item{\code{optim_hyperparameters}}{A list contanning tunable hyperparameters, check
    \code{\link{buildRNN}}.}
   \item{\code{preprocess_parameter}}{A character value indicating data preparation method,
    \code{\link{buildRNN}}.}
   \item{\code{tracking}}{A data frame containing the tracking of the gridsearch history of keras model.}
 }
}

\section{Note}{

 Check the Detail section of \code{\link{buildRNN}} to understand how SM.rnn model is created.
}

\section{Functions}{

Available functions for this class are: \itemize{
 \item \code{\link{show}}: display the information of the RNN model
 \item \code{\link{RNN_predict}}: predict function.
}
}

